requirements.txt
rembg==2.0.59
Pillow>=10.0.0
onnxruntime>=1.17.0
click>=8.1.0